+++ index.ts
@@ -1,4 +1,5 @@
 import CoT from './lib/cot.js';
+
 export * from './lib/parser.js';
 export * as Feature from './lib/types/feature.js';
 export * as Types from './lib/types/types.js';
